adding files

+++ 2025/BinhPhuoc2/BTVN.cpp
@@ -1,6 +1,8 @@
 #include <bits/stdc++.h>
 using namespace std;
 
+
+
 void sangnt(vector<int> &listnt, vector<bool> &nto)
 {
     nto[0]=nto[1]=false;
@@ -38,17 +40,17 @@ int main()
     sangnt(ntlist,nto);
     for (int i = 0; i < t; i++)
     {
-        int n;
+        long long n;
         cin>>n;
         auto pos=upper_bound(ntlist.begin(),ntlist.end(),n);
         long long acc=1;
-        for (int j = 0; j < pos-ntlist.begin(); j++)
+        for (long long j = 0; j < pos-ntlist.begin(); j++)
         {
             long long tol=0;
             long long a=mugiaithua(n,ntlist[j]);
             for (int k = 0; k <= a; k++)
-            {
-                int poww=1;
+            {   
+                long long poww=1;
                 for (int l = 0; l < k; l++)
                 {
                     poww*=ntlist[j];
